pyproject.toml
[tool.poetry]
name = "social-media-app"
version = "0.1.0"
description = ""
authors = ["kaskous <koussay.benhfaidh@esprit.tn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.0"  # Use Django 4.2.x which supports Python 3.9
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
pytest-django = "^4.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
